refactor(Money): 필드 타입을 int에서 long으로 변경

+++ src/main/java/blackjack/domain/gamer/Money.java
@@ -1,8 +1,8 @@
 package blackjack.domain.gamer;
 
-public record Money(int amount) {
+public record Money(long amount) {
 
-    private static final Money ZERO_AMOUNT = new Money(0);
+    private static final Money ZERO_AMOUNT = new Money(0L);
 
     public static Money getZeroAmountMoney() {
         return ZERO_AMOUNT;
@@ -13,6 +13,6 @@ public record Money(int amount) {
     }
 
     public Money multiplyByRatio(double ratio) {
-        return new Money((int) (amount * ratio));
+        return new Money((long) (amount * ratio));
     }
 }
