+++ src/main/java/blackjack/dto/GamerRevenueDto.java
@@ -5,10 +5,10 @@ import blackjack.domain.gamer.Money;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
-public record GamerRevenueDto(Map<String, Integer> nameRevenueMap) {
+public record GamerRevenueDto(Map<String, Long> nameRevenueMap) {
 
     public static GamerRevenueDto from(Map<BlackjackGamer, Money> gamerRevenueMap) {
-        Map<String, Integer> nameRevenueMap = new LinkedHashMap<>();
+        Map<String, Long> nameRevenueMap = new LinkedHashMap<>();
         gamerRevenueMap.forEach((gamer, revenue) -> nameRevenueMap.put(gamer.getName().value(), revenue.amount()));
 
         return new GamerRevenueDto(nameRevenueMap);
